requirements_cloud.txt
# Cloud pipeline dependencies for Supabase → BigQuery → Vertex AI

# Environment variables
python-dotenv>=1.0.0

# Supabase client
supabase>=2.0.0

# Google Cloud clients
google-cloud-bigquery>=3.11.0
google-cloud-aiplatform>=1.38.0
db-dtypes>=1.0.0

# Gemini AI
google-genai>=1.0.0

# Data processing
pandas>=2.0.0

# Optional: for better authentication handling
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1
